routes renamed to controllers, Id moved into helpers
