+++ src/controllers/friend-endpoint.js
