+++ src/controllers/friend-endpoint.spec.js
